+++ src/Weather.Domain/Dtos/LocationDto.cs
@@ -2,7 +2,7 @@
 {
     public sealed class LocationDto
     {
-        public long Latitude { get; init; }
-        public long Longitude { get; init; }
+        public double Latitude { get; init; }
+        public double Longitude { get; init; }
     }
 }
